filter button - updated styling + moved folder
