+++ components/icon/paths.ts
@@ -57,6 +57,13 @@ export const plusNormal =
 export const plusBold =
   "M28.5 16C28.5 16.3978 28.342 16.7794 28.0607 17.0607C27.7794 17.342 27.3978 17.5 27 17.5H17.5V27C17.5 27.3978 17.342 27.7794 17.0607 28.0607C16.7794 28.342 16.3978 28.5 16 28.5C15.6022 28.5 15.2206 28.342 14.9393 28.0607C14.658 27.7794 14.5 27.3978 14.5 27V17.5H5C4.60218 17.5 4.22064 17.342 3.93934 17.0607C3.65804 16.7794 3.5 16.3978 3.5 16C3.5 15.6022 3.65804 15.2206 3.93934 14.9393C4.22064 14.658 4.60218 14.5 5 14.5H14.5V5C14.5 4.60218 14.658 4.22064 14.9393 3.93934C15.2206 3.65804 15.6022 3.5 16 3.5C16.3978 3.5 16.7794 3.65804 17.0607 3.93934C17.342 4.22064 17.5 4.60218 17.5 5V14.5H27C27.3978 14.5 27.7794 14.658 28.0607 14.9393C28.342 15.2206 28.5 15.6022 28.5 16Z";
 
+export const magnifingGlassLight =
+  "M28.5298 27.47L22.1073 21.0475C23.9388 18.9413 24.881 16.2062 24.7352 13.4189C24.5895 10.6316 23.3672 8.00977 21.326 6.10608C19.2849 4.20238 16.5843 3.16556 13.7936 3.21416C11.0029 3.26276 8.34012 4.393 6.3665 6.36662C4.39288 8.34024 3.26264 11.0031 3.21404 13.7938C3.16543 16.5844 4.20226 19.285 6.10596 21.3262C8.00965 23.3673 10.6315 24.5896 13.4188 24.7353C16.2061 24.8811 18.9411 23.9389 21.0473 22.1075L27.4698 28.53C27.612 28.6624 27.8001 28.7346 27.9944 28.7311C28.1887 28.7277 28.3741 28.649 28.5115 28.5116C28.6489 28.3742 28.7276 28.1888 28.731 27.9945C28.7344 27.8002 28.6623 27.6121 28.5298 27.47ZM4.74984 14C4.74984 12.1705 5.29235 10.3821 6.30875 8.86094C7.32515 7.33979 8.7698 6.15419 10.46 5.45408C12.1502 4.75397 14.0101 4.57079 15.8044 4.9277C17.5988 5.28462 19.2469 6.16559 20.5406 7.45923C21.8342 8.75286 22.7152 10.4011 23.0721 12.1954C23.429 13.9897 23.2458 15.8496 22.5457 17.5398C21.8456 19.23 20.66 20.6747 19.1389 21.6911C17.6177 22.7075 15.8293 23.25 13.9998 23.25C11.5475 23.247 9.19646 22.2715 7.46239 20.5374C5.72833 18.8033 4.75282 16.4523 4.74984 14Z";
+export const magnifingGlassNormal =
+  "M28.7073 27.2925L22.4485 21.035C24.2626 18.8572 25.1672 16.0638 24.9741 13.236C24.781 10.4081 23.5052 7.76361 21.412 5.85251C19.3188 3.9414 16.5694 2.91086 13.7357 2.97526C10.902 3.03966 8.20225 4.19404 6.19802 6.19827C4.1938 8.20249 3.03941 10.9023 2.97501 13.7359C2.91061 16.5696 3.94116 19.319 5.85226 21.4122C7.76337 23.5054 10.4079 24.7813 13.2357 24.9743C16.0635 25.1674 18.8569 24.2628 21.0348 22.4488L27.2923 28.7075C27.3852 28.8005 27.4955 28.8742 27.6169 28.9244C27.7383 28.9747 27.8684 29.0006 27.9998 29.0006C28.1312 29.0006 28.2613 28.9747 28.3827 28.9244C28.5041 28.8742 28.6144 28.8005 28.7073 28.7075C28.8002 28.6146 28.8739 28.5043 28.9242 28.3829C28.9745 28.2615 29.0004 28.1314 29.0004 28C29.0004 27.8686 28.9745 27.7385 28.9242 27.6171C28.8739 27.4958 28.8002 27.3855 28.7073 27.2925ZM4.9998 14C4.9998 12.22 5.52764 10.48 6.51657 8.99991C7.5055 7.51987 8.91111 6.36631 10.5556 5.68513C12.2002 5.00394 14.0098 4.82571 15.7556 5.17297C17.5014 5.52024 19.1051 6.37741 20.3638 7.63608C21.6224 8.89475 22.4796 10.4984 22.8269 12.2442C23.1741 13.9901 22.9959 15.7997 22.3147 17.4442C21.6335 19.0887 20.48 20.4943 18.9999 21.4833C17.5199 22.4722 15.7798 23 13.9998 23C11.6137 22.9974 9.32601 22.0483 7.63876 20.3611C5.95151 18.6738 5.00244 16.3862 4.9998 14Z";
+export const magnifingGlassBold =
+  "M29.0611 26.9387L23.1248 21C24.9047 18.6805 25.7357 15.7709 25.4492 12.8614C25.1627 9.95181 23.7802 7.26016 21.5821 5.33244C19.3841 3.40471 16.535 2.38527 13.613 2.4809C10.6909 2.57653 7.9146 3.78008 5.84728 5.8474C3.77996 7.91472 2.57641 10.691 2.48078 13.6131C2.38514 16.5351 3.40459 19.3842 5.33231 21.5823C7.26004 23.7803 9.95169 25.1628 12.8613 25.4493C15.7708 25.7358 18.6804 24.9048 20.9998 23.125L26.9411 29.0674C27.0806 29.207 27.2463 29.3177 27.4286 29.3932C27.6109 29.4687 27.8063 29.5076 28.0036 29.5076C28.2009 29.5076 28.3963 29.4687 28.5786 29.3932C28.7609 29.3177 28.9265 29.207 29.0661 29.0674C29.2056 28.9279 29.3163 28.7623 29.3918 28.58C29.4673 28.3977 29.5062 28.2023 29.5062 28.0049C29.5062 27.8076 29.4673 27.6122 29.3918 27.4299C29.3163 27.2476 29.2056 27.082 29.0661 26.9424L29.0611 26.9387ZM5.49983 14C5.49983 12.3188 5.99835 10.6754 6.93234 9.2776C7.86633 7.87979 9.19385 6.79032 10.747 6.14698C12.3002 5.50363 14.0093 5.3353 15.6581 5.66328C17.3069 5.99125 18.8215 6.8008 20.0102 7.98954C21.199 9.17829 22.0085 10.6928 22.3365 12.3417C22.6645 13.9905 22.4961 15.6996 21.8528 17.2528C21.2095 18.8059 20.12 20.1334 18.7222 21.0674C17.3244 22.0014 15.681 22.5 13.9998 22.5C11.7462 22.4976 9.58554 21.6014 7.99198 20.0078C6.39842 18.4142 5.50215 16.2536 5.49983 14Z";
+
 export const gearLight =
   "M16.0002 10.25C14.863 10.25 13.7513 10.5872 12.8057 11.2191C11.8601 11.8509 11.1231 12.7489 10.6879 13.7996C10.2527 14.8503 10.1388 16.0064 10.3607 17.1218C10.5826 18.2372 11.1302 19.2617 11.9343 20.0659C12.7385 20.87 13.7631 21.4177 14.8784 21.6395C15.9938 21.8614 17.15 21.7475 18.2006 21.3123C19.2513 20.8771 20.1493 20.1401 20.7812 19.1945C21.413 18.249 21.7502 17.1373 21.7502 16C21.7482 14.4756 21.1418 13.0142 20.0639 11.9363C18.986 10.8584 17.5246 10.252 16.0002 10.25ZM16.0002 20.25C15.1596 20.25 14.3379 20.0008 13.639 19.5338C12.9401 19.0668 12.3954 18.403 12.0737 17.6264C11.7521 16.8498 11.6679 15.9953 11.8319 15.1709C11.9959 14.3465 12.4006 13.5892 12.995 12.9948C13.5894 12.4004 14.3467 11.9957 15.1711 11.8317C15.9955 11.6677 16.85 11.7519 17.6266 12.0735C18.4032 12.3952 19.067 12.9399 19.534 13.6388C20.001 14.3377 20.2502 15.1594 20.2502 16C20.2502 17.1272 19.8024 18.2082 19.0054 19.0052C18.2084 19.8022 17.1274 20.25 16.0002 20.25ZM26.7502 16.355C26.7577 16.1188 26.7577 15.8813 26.7502 15.645L28.6665 13.25C28.7397 13.1583 28.7904 13.0507 28.8145 12.9359C28.8386 12.8211 28.8354 12.7022 28.8052 12.5888C28.5054 11.4589 28.0565 10.3739 27.4702 9.36251C27.412 9.25987 27.3304 9.17241 27.232 9.10725C27.1336 9.04209 27.0212 9.00106 26.904 8.98751L23.8477 8.64876C23.6869 8.47626 23.5202 8.3096 23.3477 8.14876L23.0077 5.09126C22.9942 4.97403 22.9531 4.86165 22.888 4.76326C22.8228 4.66487 22.7354 4.58324 22.6327 4.52501C21.6226 3.94046 20.5394 3.49239 19.4115 3.19251C19.2978 3.16281 19.1787 3.16023 19.0639 3.18498C18.949 3.20974 18.8416 3.26113 18.7502 3.33501L16.3502 5.25626C16.114 5.24876 15.8765 5.24876 15.6402 5.25626L13.2502 3.33376C13.1585 3.26052 13.0509 3.20981 12.9361 3.18571C12.8212 3.16161 12.7024 3.1648 12.589 3.19501C11.4591 3.49483 10.3741 3.94377 9.36271 4.53001C9.26006 4.58824 9.1726 4.66987 9.10744 4.76826C9.04228 4.86665 9.00126 4.97903 8.98771 5.09626L8.65396 8.15251C8.48146 8.31418 8.31479 8.48085 8.15396 8.65251L5.09521 9.00001C4.97798 9.01356 4.8656 9.05459 4.76721 9.11975C4.66881 9.18491 4.58719 9.27237 4.52896 9.37501C3.94441 10.3852 3.49634 11.4684 3.19646 12.5963C3.16747 12.7085 3.16492 12.826 3.18899 12.9394C3.21307 13.0529 3.26312 13.1592 3.33521 13.25L5.25646 15.65C5.24896 15.8863 5.24896 16.1238 5.25646 16.36L3.33396 18.7563C3.26072 18.8479 3.21 18.9555 3.18591 19.0704C3.16181 19.1852 3.165 19.3041 3.19521 19.4175C3.49555 20.5453 3.94448 21.6281 4.53021 22.6375C4.58844 22.7402 4.67006 22.8276 4.76846 22.8928C4.86685 22.9579 4.97923 22.999 5.09646 23.0125L8.15271 23.3513C8.31438 23.5238 8.48104 23.6904 8.65271 23.8513L9.00021 26.905C9.01376 27.0222 9.05478 27.1346 9.11994 27.233C9.1851 27.3314 9.27256 27.413 9.37521 27.4713C10.3854 28.0558 11.4685 28.5039 12.5965 28.8038C12.7099 28.834 12.8287 28.8372 12.9436 28.8131C13.0584 28.789 13.166 28.7383 13.2577 28.665L15.6452 26.75C15.8815 26.7575 16.119 26.7575 16.3552 26.75L18.7565 28.6725C18.8896 28.7787 19.0549 28.8365 19.2252 28.8363C19.2902 28.836 19.3548 28.8276 19.4177 28.8113C20.5455 28.5109 21.6283 28.062 22.6377 27.4763C22.7404 27.418 22.8278 27.3364 22.893 27.238C22.9581 27.1396 22.9992 27.0272 23.0127 26.91L23.3515 23.8538C23.524 23.6929 23.6906 23.5263 23.8515 23.3538L26.909 23.0138C27.0262 23.0002 27.1386 22.9592 27.237 22.894C27.3354 22.8289 27.417 22.7414 27.4752 22.6388C28.06 21.6287 28.5081 20.5455 28.8077 19.4175C28.8379 19.3041 28.8411 19.1852 28.817 19.0704C28.7929 18.9555 28.7422 18.8479 28.669 18.7563L26.7502 16.355ZM26.3627 21.5588L23.4077 21.8875C23.2231 21.9088 23.0529 21.9979 22.9302 22.1375C22.6832 22.4158 22.4198 22.6792 22.1415 22.9263C22.0018 23.049 21.9127 23.2191 21.8915 23.4038L21.5627 26.3575C20.8709 26.7265 20.1439 27.0254 19.3927 27.25L17.0715 25.3925C16.9383 25.2863 16.773 25.2285 16.6027 25.2288H16.5577C16.1863 25.25 15.8141 25.25 15.4427 25.2288C15.2572 25.2174 15.0742 25.2753 14.929 25.3913L12.609 27.25C11.8584 27.0242 11.1323 26.724 10.4415 26.3538L10.1127 23.4013C10.0914 23.2166 10.0023 23.0465 9.86271 22.9238C9.5844 22.6767 9.32102 22.4133 9.07396 22.135C8.95127 21.9954 8.78111 21.9063 8.59646 21.885L5.64271 21.5563C5.27415 20.8664 4.97517 20.1416 4.75021 19.3925L6.60771 17.0713C6.72365 16.9261 6.78156 16.743 6.77021 16.5575C6.74896 16.1862 6.74896 15.8139 6.77021 15.4425C6.78156 15.257 6.72365 15.074 6.60771 14.9288L4.75021 12.6088C4.97602 11.8582 5.27626 11.1321 5.64646 10.4413L8.59896 10.1125C8.78361 10.0912 8.95377 10.0021 9.07646 9.86251C9.32352 9.5842 9.5869 9.32082 9.86521 9.07376C10.0048 8.95107 10.0939 8.78091 10.1152 8.59626L10.444 5.64251C11.1338 5.27395 11.8586 4.97498 12.6077 4.75001L14.929 6.60751C15.0742 6.72345 15.2572 6.78136 15.4427 6.77001C15.8141 6.74876 16.1863 6.74876 16.5577 6.77001C16.7432 6.78136 16.9263 6.72345 17.0715 6.60751L19.3915 4.75001C20.142 4.97583 20.8681 5.27607 21.559 5.64626L21.8877 8.60126C21.909 8.78591 21.9981 8.95607 22.1377 9.07876C22.416 9.32582 22.6794 9.5892 22.9265 9.86751C23.0492 10.0071 23.2193 10.0962 23.404 10.1175L26.3577 10.4463C26.7261 11.1353 27.0251 11.8593 27.2502 12.6075L25.3927 14.9288C25.2768 15.074 25.2189 15.257 25.2302 15.4425C25.2515 15.8139 25.2515 16.1862 25.2302 16.5575C25.2189 16.743 25.2768 16.9261 25.3927 17.0713L27.2502 19.3913C27.0252 20.1417 26.7258 20.8678 26.3565 21.5588H26.3627Z";
 export const gearNormal =
@@ -91,6 +98,14 @@ export const funnelNormal =
   "M28.825 6.19122C28.6711 5.83561 28.416 5.53309 28.0915 5.32129C27.767 5.10949 27.3875 4.99775 27 4.99997H4.99997C4.61288 5.00073 4.23433 5.11381 3.91025 5.32548C3.58616 5.53715 3.33047 5.83832 3.17418 6.19245C3.01789 6.54658 2.96772 6.93846 3.02977 7.32054C3.09181 7.70262 3.2634 8.05849 3.52372 8.34497L3.53372 8.35622L12 17.3962V27C11.9999 27.3619 12.098 27.7172 12.284 28.0277C12.4699 28.3383 12.7366 28.5926 13.0557 28.7635C13.3748 28.9344 13.7343 29.0155 14.0958 28.9981C14.4574 28.9808 14.8075 28.8656 15.1087 28.665L19.1087 25.9975C19.3829 25.8148 19.6078 25.5673 19.7632 25.2768C19.9187 24.9863 20 24.6619 20 24.3325V17.3962L28.4675 8.35622L28.4775 8.34497C28.7405 8.0598 28.9138 7.70346 28.9756 7.32043C29.0374 6.93741 28.985 6.54466 28.825 6.19122ZM18.2725 16.3225C18.0995 16.5059 18.0021 16.7479 18 17V24.3325L14 27V17C14 16.746 13.9035 16.5016 13.73 16.3162L4.99997 6.99997H27L18.2725 16.3225Z";
 export const funnelBold =
   "M29.2863 5.98875C29.0903 5.54581 28.77 5.16931 28.3641 4.90502C27.9582 4.64072 27.4843 4.50002 27 4.5H5C4.51575 4.50005 4.04193 4.64074 3.63611 4.90497C3.2303 5.16921 2.90997 5.54561 2.71403 5.98846C2.51809 6.43131 2.45499 6.92153 2.53238 7.39956C2.60978 7.87759 2.82434 8.32285 3.15 8.68125L3.165 8.69875L11.5 17.5938V27C11.4999 27.4526 11.6227 27.8967 11.8553 28.285C12.0879 28.6733 12.4215 28.9912 12.8206 29.2047C13.2197 29.4182 13.6692 29.5194 14.1213 29.4974C14.5734 29.4755 15.011 29.3312 15.3875 29.08L19.3875 26.4137C19.73 26.1853 20.0107 25.8757 20.2048 25.5127C20.3989 25.1496 20.5003 24.7442 20.5 24.3325V17.5938L28.8338 8.69875L28.8488 8.68125C29.1746 8.32304 29.3894 7.87791 29.4671 7.39993C29.5448 6.92195 29.4819 6.4317 29.2863 5.98875ZM17.9113 15.975C17.6488 16.2519 17.5017 16.6185 17.5 17V24.065L14.5 26.065V17C14.4996 16.6191 14.3544 16.2527 14.0938 15.975L6.15375 7.5H25.8463L17.9113 15.975Z";
+
+export const sortAscLight = "";
+export const sortAscNormal = "";
+export const sortAscBold = "";
+
+export const sortDescLight = "";
+export const sortDescNormal = "";
+export const sortDescBold = "";
 
 /* ===== ARROWS - CARET - CHEVRON ICONS ===== */
 export const caretUpLight =
